Adds component for map features
