+++ index.html
@@ -29,7 +29,7 @@
     </header>
     <section class="squirrel-map">
       <div class="map-wrapper">
-          <div id="mapid"></div>
+          <div id="mapid" component="map"></div>
           <div>Map squirrel icon made by <a href="https://www.flaticon.com/authors/eucalyp" title="Eucalyp">Eucalyp</a> from <a
               href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a></div>
       </div>
